Add files via upload
Breast cancer is a very common disease around the world and the second leading
cause of cancer death in women. Expert systems are developed using data mining methods in
order for disease diagnosis and significant tools for assisting medical doctors in their clinical
decision. In this study, k-Nearest Neighbors algorithm (k-NN) was used in order to classify
breast cancer disease. Besides, k-NN was implemented for different k values and the obtained
classification accuracies were compared with each other. According to the study results, it was
seen that breast cancer disease was successfully classified using k-NN.