Add files via upload
You've been given a classified data set from a company! They've hidden the feature column names but have given you the data and the target classes.

I will try to use KNN to create a model that directly predicts a class for a new data point based off of the features.