Add files via upload
The purpose of this analysis is to build a prediction model to predict whether a review on the restaurant is positive or negative. To do so, I will work on Restaurant Review dataset, I will load it into predicitve algorithms like  Naive Bayes and Logistic Regression  to find a "best" model for predicting the review's sentiment.

Dataset: Restaurant_Reviews.tsv is a dataset from Kaggle datasets which consists of 1000 reviews on a restaurant.

To build a model to predict if review is positive or negative, following steps are performed.